Create BMIinTheNBA
I'm a lifelong NBA fan. I wanted to explore BMI as it relates to the length of a player's career in the NBA. Does having a heavier frame or a larger BMI have a negative effect on the length of a player's NBA career. I also look at other different factors such as height and the era in which a player started. I end up with a kmeans cluster of era, height, BMI and length of career. Thanks.